implement dummy api for transfer money in and out

+++ src/main/java/com/example/virtualwallet/services/contracts/TransferService.java
@@ -2,6 +2,7 @@ package com.example.virtualwallet.services.contracts;
 
 import com.example.virtualwallet.models.Transfer;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface TransferService {
@@ -16,9 +17,9 @@ public interface TransferService {
 
     void deleteTransfer(Transfer transfer);
 
-    void transferMoney(int senderUserId, int recipientUserId, int walletId, double amount);
+    void transferMoney(int senderUserId, int recipientUserId, int walletId, BigDecimal amount);
 
     void confirmTransfer(int transferId, int senderUserId, int recipientId, int recipientWalletId);
 
-    void editTransfer(int transferId, double newAmount);
+    void editTransfer(int transferId, BigDecimal newAmount);
 }
